requirements.txt
# local package
# -e .

# external requirements
click==8.0.4
Sphinx
coverage==6.3.2
awscli==1.22.73
flake8
python-dotenv>=0.5.1
setuptools==60.9.3
pandas~=1.4.1
matplotlib~=3.5.1
plotly
numpy==1.22.3
scikit-learn~=1.0.2
imblearn~=0.0
xgboost~=1.5.2
lightgbm~=3.3.2
seaborn~=0.11.2